Add explicit type cast to `Summary`

+++ AdventOfCode2023/tasks/Day13.cs
@@ -85,7 +85,7 @@ public class AshPattern
         get
         {
             _summary ??= GetSummary();
-            return _summary;
+            return (int)_summary;
         }
     }
